requirements.txt
apache-airflow==2.10.5
pandas==2.0.3
fhir.resources==7.1.0